% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shellcheckr.R
\name{shellcheckr}
\alias{shellcheckr}
\title{Run \code{shellcheck} on a document}
\usage{
shellcheckr(path)
}
\arguments{
\item{path}{path of file to check. The active document is used if this is missing.}
}
\value{
JSON output of \href{https://www.shellcheck.net/}{\code{shellcheck}}.
}
\description{
Runs \href{https://www.shellcheck.net/}{\code{shellcheck}} on a document and returns the JSON output.
}
