% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shellcheckrAddin.R
\name{shellcheckrAddin}
\alias{shellcheckrAddin}
\title{Run \code{shellcheck} on a document}
\usage{
shellcheckrAddin(path)
}
\arguments{
\item{path}{path of file to check. The active document is used if this is missing.}
}
\description{
A function which parses \href{https://www.shellcheck.net/}{\code{shellcheck}} output into markers with \href{https://www.shellcheck.net/wiki/}{ShellCheck wiki} links.
Intended for use as an \href{https://rstudio.github.io/rstudioaddins/}{RStudio Addin}.
}
