% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shellcheckr.R
\name{shellcheckrMarkers}
\alias{shellcheckrMarkers}
\title{Run \code{shellcheck} on a document and parse to markers}
\usage{
shellcheckrMarkers(path)
}
\arguments{
\item{path}{path of file to check. The active document is used if this is missing.}
}
\description{
Runs \href{https://www.shellcheck.net/}{\code{shellcheck}} and parses output into markers with
\href{https://www.shellcheck.net/wiki/}{ShellCheck wiki} links for the problematic codes.
}
\details{
Intended for use as an \href{https://rstudio.github.io/rstudioaddins/}{RStudio Addin}.
}
